+++ src/page-components/retros/NewRetroModal.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect, useRef } from "react";
 import BaseIcon from "@/components/BaseIcon";
 import { ICONS } from "@/helpers/constant";
-import boardService from "@/services/board.service";
 import { useParams } from "react-router-dom";
 import { toast } from "react-toastify";
 import BaseButton from "@/components/BaseButton";
 import BaseInput from "@/components/form-inputs/BaseInput";
 import BaseModal from "@/components/BaseModal";
+import retroService from "@/services/retro.service";
 
 export default function BaseForm(props) {
   const { children } = props;
@@ -49,7 +49,7 @@ export default function BaseForm(props) {
   }, [isOpen]);
 
   const handleSave = async () => {
-    await boardService.createRetro({ boardId: params.boardId }, retroModel);
+    await retroService.createRetro({ boardId: params.boardId }, retroModel);
     closeModal();
     props.getBoardRetros();
     toast.success("Retro created successfully.");
